code optimization with templates, static arrays and asynchronous copy
1) The code has been sped up considerably by introducing templated kernel functions.
For each data type (float or double) and each combination of angular momenta of the
basis functions and the type of polarization operator a specialized kernel is compiled.
The compiler knows the limits of many loops and can unroll them.

2) All dynamic memory allocations inside the kernel were replaced by static arrays,
since the array boundaries are now constant template parameters.

3) Kernel computations and copy operations from and to the device are overlapped
by using different streams and non-blocking cudaMemcpyAsync. Allocation of the
pinned CPU memory is a huge bottleneck.